# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ГОСТИННИЦЫ - API",
        "description": "Описание всех методов приложения ГОСТИННИЦЫ",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/rooms/get": {
            "get": {
                "tags": [
                    "Комнаты"
                ],
                "summary": "Список комнат",
                "description": "Список комнат",
                "operationId": "indexRooms",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": {
                                    "filter": {
                                        "priceMin": "0",
                                        "priceMax": "0",
                                        "capacityMin": "0",
                                        "capacityMax": "0"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "hotels_id": "1",
                                    "number": "1",
                                    "capacity": "1",
                                    "price": "42.23",
                                    "description": "Комната",
                                    "deleted_at": "2023-07-12 06:45:23",
                                    "created_at": "2023-07-12 06:45:23",
                                    "updated_at": "2023-07-12 06:45:23"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/rooms/getById": {
            "get": {
                "tags": [
                    "Комнаты"
                ],
                "summary": "Получить объект комнаты по ID",
                "description": "Получить объект комнаты по ID",
                "operationId": "showRoom",
                "parameters": [
                    {
                        "name": "room_id",
                        "in": "query",
                        "description": "ID комнаты",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "hotels_id": "1",
                                    "number": "1",
                                    "capacity": "1",
                                    "price": "42.23",
                                    "description": "Комната",
                                    "deleted_at": "2023-07-12 06:45:23",
                                    "created_at": "2023-07-12 06:45:23",
                                    "updated_at": "2023-07-12 06:45:23"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/rooms/getByHotelId": {
            "get": {
                "tags": [
                    "Комнаты"
                ],
                "summary": "Получить объект комнаты по ID отеля",
                "description": "Получить объект комнаты по ID отеля",
                "operationId": "showRoomByHotelId",
                "parameters": [
                    {
                        "name": "hotel_id",
                        "in": "query",
                        "description": "ID комнаты",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "hotels_id": "1",
                                    "number": "1",
                                    "capacity": "1",
                                    "price": "42.23",
                                    "description": "Комната",
                                    "deleted_at": "2023-07-12 06:45:23",
                                    "created_at": "2023-07-12 06:45:23",
                                    "updated_at": "2023-07-12 06:45:23"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/rooms/store": {
            "post": {
                "tags": [
                    "Комнаты"
                ],
                "summary": "Создание объекта отеля",
                "description": "Создание объекта отеля",
                "operationId": "storeRoom",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": {
                                    "hotels_id": "1",
                                    "number": "1",
                                    "capacity": "1",
                                    "price": "42.23",
                                    "description": "Комната",
                                    "deleted_at": "2023-07-12 06:45:23",
                                    "created_at": "2023-07-12 06:45:23",
                                    "updated_at": "2023-07-12 06:45:23"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успех",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "hotels_id": "1",
                                        "number": "1",
                                        "capacity": "1",
                                        "price": "42.23",
                                        "description": "Комната",
                                        "deleted_at": "2023-07-12 06:45:23",
                                        "created_at": "2023-07-12 06:45:23",
                                        "updated_at": "2023-07-12 06:45:23"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/rooms/update": {
            "put": {
                "tags": [
                    "Комнаты"
                ],
                "summary": "Обновление объекта комнаты",
                "description": "Обновление объекта комнаты",
                "operationId": "updateRoom",
                "parameters": [
                    {
                        "name": "room_id",
                        "in": "query",
                        "description": "ID комнаты",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": {
                                    "hotels_id": "1",
                                    "number": "1",
                                    "capacity": "1",
                                    "price": "42.23",
                                    "description": "Комната",
                                    "deleted_at": "2023-07-12 06:45:23",
                                    "created_at": "2023-07-12 06:45:23",
                                    "updated_at": "2023-07-12 06:45:23"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успех",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "hotels_id": "1",
                                        "number": "1",
                                        "capacity": "1",
                                        "price": "42.23",
                                        "description": "Комната",
                                        "deleted_at": "2023-07-12 06:45:23",
                                        "created_at": "2023-07-12 06:45:23",
                                        "updated_at": "2023-07-12 06:45:23"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/rooms/delete": {
            "delete": {
                "tags": [
                    "Комнаты"
                ],
                "summary": "Удаление объекта комнаты по ID",
                "description": "Удаление объекта комнаты по ID",
                "operationId": "destroyRoom",
                "parameters": [
                    {
                        "name": "room_id",
                        "in": "query",
                        "description": "ID комнаты",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        }
    }
}